Fix markup

+++ tfm.api.dal/Repos/Implemetations/StylePriceRepo.cs
@@ -2,6 +2,7 @@
 using tfm.api.dal.Db;
 using tfm.api.dal.Entities;
 using tfm.api.dal.Repos.Contracts;
+using tfm.api.exceptions;
 
 namespace tfm.api.dal.Repos.Implemetations
 {
@@ -14,7 +15,7 @@ namespace tfm.api.dal.Repos.Implemetations
             _db = context;
         }
 
-        public async Task<int> AddAsync(StylePrice stylePrice)
+        public async Task<int> AddAsync(StylePriceEntity stylePrice)
         {
             if (stylePrice is null)
             {
@@ -28,14 +29,23 @@ namespace tfm.api.dal.Repos.Implemetations
             return stylePrice.Id;
         }
 
-        public async Task DeleteAsync(int stylePriceid)
+        public async Task DeleteAsync(int stylePriceId)
         {
-            StylePrice? entity = await _db.StylePrices.FirstOrDefaultAsync(_ => _.Id == stylePriceid)
-                                ?? throw new NotFoundException($"StylePriceId = {stylePriceid}. Can't find specified item.");
+            if (stylePriceId <= 0) throw new ArgumentOutOfRangeException(nameof(stylePriceId));
+            
+            StylePriceEntity? entity = await _db.StylePrices.FirstOrDefaultAsync(_ => _.Id == stylePriceId)
+                                       ?? throw new NotFoundException($"StylePriceId = {stylePriceId}. Can't find specified item.");
 
             _db.StylePrices.Remove(entity);
 
             await _db.SaveChangesAsync();
+        }
+
+        public async Task<bool> IsExistAsync(int masterId, int styleId)
+        {
+            if (masterId <= 0) throw new ArgumentOutOfRangeException(nameof(masterId));
+            
+            return await _db.StylePrices.AnyAsync(_ => _.MasterId == masterId && _.StyleId == styleId);
         }
     }
 }
